ButtonV2 code cleanup
