+++ packages/button/src/themes.ts
@@ -1,0 +1,35 @@
+import { colors } from '@ndla/core';
+import { ButtonColor, ButtonTheme } from './types';
+
+export const themes: Record<ButtonColor, ButtonTheme> = {
+  primary: {
+    foreground: colors.white,
+    background: colors.brand.primary,
+    hoverForeground: colors.white,
+    hoverBackground: colors.brand.dark,
+  },
+  light: {
+    foreground: colors.brand.primary,
+    background: colors.brand.light,
+    hoverForeground: colors.white,
+    hoverBackground: colors.brand.dark,
+  },
+  lighter: {
+    foreground: colors.brand.primary,
+    background: colors.brand.lighter,
+    hoverForeground: colors.white,
+    hoverBackground: colors.brand.primary,
+  },
+  greyLighter: {
+    foreground: colors.brand.primary,
+    background: colors.brand.greyLighter,
+    hoverForeground: colors.white,
+    hoverBackground: colors.brand.primary,
+  },
+  greyLightest: {
+    foreground: colors.brand.primary,
+    background: colors.brand.greyLightest,
+    hoverForeground: colors.white,
+    hoverBackground: colors.brand.primary,
+  },
+};
